# Conflicts:
#	blackboard2/target/classes/application.properties

+++ blackboard2/src/main/java/com/example/demo/Service/CourseService.java
@@ -11,7 +11,7 @@ public interface CourseService {
 
 		Course findByCourseCode(String code);
 		Course findByUuid(String code);
-		List<Course> findByTeacherId(long id);
+		List<Course> findByTeacherId(int id);
 		List<Course> findAll();
 
 }
